Restructure bot
Remove src folder from bot
Rename standard.Dockerfile to Dockerfile
Remove standard.Dockerfile from compose.yaml
