+++ commands/ping.ts
