+++ commands/pun.ts
