+++ handlers/command.ts
@@ -2,7 +2,7 @@ import { Client, Routes, SlashCommandBuilder } from 'discord.js';
 import { REST } from '@discordjs/rest';
 import { readdirSync } from 'fs';
 import { join } from 'path';
-import { color } from '../functions';
+import { color } from '../utils/functions';
 import { SlashCommand } from '../types';
 
 module.exports = (client: Client) => {
