--- src/events/ready.ts
@@ -1,15 +1,0 @@
-import { Client } from "discord.js";
-import { BotEvent } from "../types";
-import { color } from "../functions";
-
-const event: BotEvent = {
-  name: "ready",
-  once: true,
-  execute: (client: Client) => {
-    console.log(
-      color("text", `💪 Logged in as ${color("variable", client.user?.tag)}`)
-    );
-  },
-};
-
-export default event;
