+++ tsconfig.json
@@ -26,7 +26,7 @@
 
 		/* Modules */
 		"module": "CommonJS" /* Specify what module code is generated. */,
-		"rootDir": "./src" /* Specify the root folder within your source files. */,
+		"rootDir": "./" /* Specify the root folder within your source files. */,
 		// "moduleResolution": "node",                       /* Specify how TypeScript looks up a file from a given module specifier. */
 		// "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
 		// "paths": {},                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
@@ -100,5 +100,6 @@
 		// "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
 		"skipLibCheck": true /* Skip type checking all .d.ts files. */
 	},
-	"include": ["./src/**/*", "src/commands/ping.ts", "src/commands/ping.ts"]
+	"include": ["./", "commands/ping.ts", "commands/ping.ts"],
+	"exclude": ["./build", "./node_modules", "./registerCommands"]
 }
